move this ugly mess to separate file for now...
